feat: Remove NewsExtensions portlet from bottom-all-container configuration - EXO-69984
Priori to this change, in bottom-all-container, we inject and load news
extensions as a portlet, so, it is loaded on all pages even if no need
to be used which can lead to loading performance issues.
To fix that, we must use js load-group, which are now available through
all webapps to load news extensions only with the portlets of the same
load-group (example activity stream, notification, analytics table,
favorite drawer).
