+++ webapp/src/main/webapp/news-extensions/activity-stream-extensions/components/activity/ActivitySwitchToNews.vue
