+++ webapp/src/main/webapp/news-extensions/activity-stream-extensions/components/activity/ActivityWriteNewsComposer.vue
