+++ webapp/src/main/webapp/news-extensions/activity-stream-extensions/extensions.js
@@ -13,6 +13,13 @@
   You should have received a copy of the GNU Affero General Public License
   along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
+
+import {canUserCreateNews} from '../../services/newsServices.js';
+const lang = eXo.env.portal.language || 'en';
+const url = `${eXo.env.portal.context}/${eXo.env.portal.rest}/i18n/bundle/locale.portlet.news.News-${lang}.json`;
+
+const i18nPromise = exoi18n.loadLanguageAsync(lang, url).then(i18n => new Vue({i18n}));
+
 const newsActivityTypeExtensionOptions = {
   name: 'News',
   getExtendedComponent: (activity, isActivityDetail) => {
@@ -73,4 +80,27 @@ export function initExtensions() {
     type: 'news',
     options: newsActivityTypeExtensionOptions,
   });
+  if (eXo.env.portal.spaceId) {
+    canUserCreateNews(eXo.env.portal.spaceId).then(canCreateNews => {
+      if (canCreateNews) {
+        return i18nPromise.then(() => {
+          extensionRegistry.registerComponent('ActivityComposerAction', 'activity-composer-action', {
+            id: 'switchNewsButton',
+            vueComponent: Vue.options.components['activity-switch-to-news'],
+            rank: 10,
+          });
+          extensionRegistry.registerComponent('ActivityComposerFooterAction', 'activity-composer-footer-action', {
+            id: 'writeNewsButton',
+            vueComponent: Vue.options.components['activity-write-news-composer'],
+            rank: 30,
+          });
+          extensionRegistry.registerComponent('ActivityToolbarAction', 'activity-toolbar-action', {
+            id: 'writeNewsToolbarButton',
+            vueComponent: Vue.options.components['activity-write-news-toolbar-action'],
+            rank: 30,
+          });
+        });
+      }
+    });
+  }
 }
