+++ webapp/src/main/webapp/news-extensions/activity-stream-extensions/initComponents.js
