+++ webapp/src/main/webapp/news-extensions/activity-stream-extensions/main.js
@@ -13,7 +13,15 @@
   You should have received a copy of the GNU Affero General Public License
   along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
+import './initComponents.js';
 import {initExtensions} from './extensions.js';
+import * as  newsServices from '../../services/newsServices';
+
+if (!Vue.prototype.$newsServices) {
+  window.Object.defineProperty(Vue.prototype, '$newsServices', {
+    value: newsServices,
+  });
+}
 
 export function init() {
   initExtensions();
