+++ webapp/src/main/webapp/news-extensions/notification-extensions/components/PostNewsNotificationPlugin.vue
