+++ webapp/src/main/webapp/news-extensions/notification-extensions/extensions.js
