+++ webapp/src/main/webapp/news-extensions/notification-extensions/initComponents.js
