+++ webapp/src/main/webapp/news-extensions/notification-extensions/main.js
