+++ webapp/webpack.common.js
@@ -13,15 +13,14 @@ let config = {
     latestNews:'./src/main/webapp/latest-news/main.js',
     news :'./src/main/webapp/news/main.js',
     newsSearchCard: './src/main/webapp/news-search/main.js',
-    newsExtensions: './src/main/webapp/news-extensions/main.js',
     newsDetailsApp: './src/main/webapp/news-details-app/main.js',
     scheduleNewsDrawer: './src/main/webapp/schedule-news-drawer/main.js',
     newsListView: './src/main/webapp/news-list-view/main.js',
     newsPublishTargetsManagement: './src/main/webapp/news-publish-targets-management/main.js',
-    newsActivityExtensions: './src/main/webapp/news-extensions/activity-extensions/main.js',
-    newsComposerActionExtensions: './src/main/webapp/news-extensions/composer-action-extensions/main.js',
+    newsActivityStreamExtensions: './src/main/webapp/news-extensions/activity-stream-extensions/main.js',
     newsFavoriteDrawerExtensions: './src/main/webapp/news-extensions/favorite-drawer-extensions/main.js',
-    newsNotificationsExtension: './src/main/webapp/news-notifications/main.js',
+    newsAnalyticsExtensions: './src/main/webapp/news-extensions/analytics-extensions/main.js',
+    newsNotificationExtensions: './src/main/webapp/news-extensions/notification-extensions/main.js',
   },
   output: {
     filename: 'js/[name].bundle.js',
